moved temperature conversion methods into utils class
