+++ client/src/WeatherDashboard.js
@@ -3,6 +3,7 @@ import axios from 'axios';
 import 'assets/css/weatherDashboard.css';
 import WeatherList from 'components/WeatherList';
 import ToggleAddLocation from 'components/ToggleAddLocation';
+import Convert from 'util/convertUtil';
 
 class WeatherDashboard extends Component {
 
@@ -43,9 +44,9 @@ class WeatherDashboard extends Component {
 					data.sys.sunset = this.convertDateTime(data.sys.sunset);
 
 					// convert init temp to celcius
-					data.main.temp = this.convertKelvinToCelcius(res.data.main.temp);
-					data.main.temp_min =this.convertKelvinToCelcius(res.data.main.temp_min);
-					data.main.temp_max = this.convertKelvinToCelcius(res.data.main.temp_max);
+					data.main.temp = this.initTempConversion(res.data.main.temp);
+					data.main.temp_min =this.initTempConversion(res.data.main.temp_min);
+					data.main.temp_max = this.initTempConversion(res.data.main.temp_max);
 
 					return this.setState({ locations: this.state.locations.concat(data), lat: null, long: null });//  res.data;
 				})
@@ -57,13 +58,11 @@ class WeatherDashboard extends Component {
  
 
 	convertDateTime = (timestamp) => {
-		let d = new Date(timestamp*1000);
-		let t = d.toTimeString().split(' ')[0];
-		return t;
+		return Convert.convertDateTime(timestamp);
 	}
 
-	convertKelvinToCelcius = (temp) => {
-		return Math.round(temp) -270;
+	initTempConversion = (temp) => {
+		return Convert.initTempConversion(temp);
 	}
 
 
@@ -84,59 +83,17 @@ class WeatherDashboard extends Component {
 
 
 	removeLocation = (city) => {
-
 		this.setState({
 			locations: this.state.locations.filter( loc => loc.name !== city )
 		})
-
 	}
 
-	conversion = (temps, degree, city) => {	
-
-		let convertTempCalc;
-		let updatedTempData = [];
-		//let convertedLocationObject;
-		let updatedDegree;
-
-		temps.forEach((temp) => {
-
-			if(degree === 'C'){
-				convertTempCalc = (9.0/5.0 * temp) + 32;
-				updatedDegree = 'F';
-
-			} else {
-				convertTempCalc = 5.0/9.0 * (temp-32);
-				updatedDegree = 'C';
-			}
-			updatedTempData.push(Math.round(convertTempCalc))
-		})
-
-		const temperatureArray = this.state.locations.map((loc) => {
-		
-			if(loc.name === city){
-				
-				let convertedTempObject = Object.assign({}, loc.main, {
-					temp: updatedTempData[0],
-					temp_max: updatedTempData[1],
-					temp_min: updatedTempData[2]
-				})
-
-				return Object.assign({}, loc, {
-					main: convertedTempObject,
-					degree: updatedDegree
-				})
-
-			} else {
-				return loc;
-			}
-			
-		})
+	convertTemperatures = (temps, degree, city) => {	
+		const temperatureArray = Convert.convertTemperatures(temps, degree, city, this.state);
 
 		this.setState({
 			locations: temperatureArray
 		})
-			
-		console.log('set state', temperatureArray)
 	}
 
 
@@ -147,7 +104,7 @@ class WeatherDashboard extends Component {
 			   	
 				   	<WeatherList 
 				   		locations = {this.state.locations}
-				   		onConversion = {this.conversion}
+				   		onConversion = {this.convertTemperatures}
 				   		onRemove = {this.removeLocation}
 				   	/>
